[WIP] fix parsing error
