+++ packages/core/package.json
@@ -10,6 +10,7 @@
   "license": "Apache-2.0",
   "dependencies": {
     "bean-parser": "^1.0.9",
-    "source-map": "^0.7.3"
+    "source-map": "^0.7.3",
+    "clio-highlight": "^1.0.0"
   }
 }
