+++ packages/core/parser.js
@@ -1,6 +1,7 @@
 const { bean } = require("bean-parser");
 const lexer = require("./lexer.js");
 const model = require("./model.js");
+const { colorize } = require("clio-highlight");
 
 const parse = tokens => bean(model, tokens);
 
@@ -17,8 +18,19 @@ const parser = source =>
           .filter(m => m.left == firstUnmatched)
           .map(({ right }) => right);
         const encountered = result[1].name;
-        const { index } = result[1];
-        const ParsingError = `At ${index}: Expecting one of ${expecting.join(
+        const { location } = result[1];
+        const { line, column } = location;
+        const lines = source.split("\n").slice(Math.max(0, line - 3), line);
+        const lineCount = lines.length;
+        console.log();
+        lines.forEach((rawLine, index) => {
+          const lineNumber = line - lineCount + index;
+          const padding = " ".repeat((line - lineNumber).toString().length + 1);
+          const coloredLine = colorize(rawLine);
+          console.log(`${padding}${lineNumber} | ${coloredLine}`);
+        });
+        console.log(" ".repeat(column + 4 + line.toString().length - 1) + "^");
+        const ParsingError = `At ${line},${column}: Expecting one of ${expecting.join(
           ", "
         )} but encountered ${encountered}`;
         throw ParsingError;
