+++ packages/highlight/package.json
@@ -1,0 +1,15 @@
+{
+  "name": "clio-highlight",
+  "version": "1.0.0",
+  "description": "",
+  "main": "index.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "keywords": [],
+  "author": "Pouya Eghbali <pouya.eghbali@yandex.com>",
+  "license": "Apache-2.0",
+  "dependencies": {
+    "chalk": "3.0.0"
+  }
+}
